🔧 FIX: Cast days_until_expiry to int in CreateExpiredTestUsers command
- Fixed PostgreSQL error: invalid input syntax for type integer
- Production database requires strict int type for days_until_expiry
- Cast Carbon diffInDays result to (int) for PostgreSQL compatibility
- Now production expired users setup will work properly

+++ app/Console/Commands/CreateExpiredTestUsers.php
@@ -83,7 +83,7 @@ class CreateExpiredTestUsers extends Command
             // Calculate membership dates
             $membershipStartDate = Carbon::now()->subYear(); // Started a year ago
             $membershipEndDate = Carbon::now()->subDays($userData['expired_days_ago']); // Expired X days ago
-            $daysUntilExpiry = Carbon::now()->diffInDays($membershipEndDate, false); // Negative for expired
+            $daysUntilExpiry = (int) Carbon::now()->diffInDays($membershipEndDate, false); // Negative for expired (cast to int)
 
             // Create membership renewal record
             $membershipRenewal = MembershipRenewal::create([
